backend/docs/markdown file/tests/test_vision_service.py
"""
Tests for vision service crop disease identification.
"""

import pytest
import asyncio
import base64
import io
from unittest.mock import Mock, patch, AsyncMock
from PIL import Image

from app.services.vision_service import (
    vision_service,
    VisionService,
    ImageValidationError,
    DiseaseIdentificationError,
    DiseaseConfidence,
    ImageFormat,
)


class TestVisionService:
    """Test cases for VisionService."""

    def setup_method(self):
        """Set up test fixtures."""
        self.vision_service = VisionService()

    def create_test_image(self, format="JPEG", size=(100, 100), color="green"):
        """Create a test image for testing."""
        image = Image.new("RGB", size, color)
        buffer = io.BytesIO()
        image.save(buffer, format=format)
        return buffer.getvalue()

    def test_validate_image_success(self):
        """Test successful image validation."""
        # Create a valid test image
        image_data = self.create_test_image()

        result = self.vision_service.validate_image(image_data)

        assert result.is_valid is True
        assert result.error_message is None
        assert result.format == "jpeg"
        assert result.size == (100, 100)
        assert result.file_size_mb is not None

    def test_validate_image_too_large(self):
        """Test image validation with oversized file."""
        # Create a large image (simulate by creating large data)
        large_data = b"x" * (11 * 1024 * 1024)  # 11MB

        result = self.vision_service.validate_image(large_data)

        assert result.is_valid is False
        assert "too large" in result.error_message.lower()

    def test_validate_image_invalid_format(self):
        """Test image validation with invalid format."""
        # Invalid image data
        invalid_data = b"not an image"

        result = self.vision_service.validate_image(invalid_data)

        assert result.is_valid is False
        assert result.error_message is not None

    def test_preprocess_image(self):
        """Test image preprocessing."""
        # Create a test image
        original_data = self.create_test_image(size=(2000, 2000))  # Large image

        processed_data = self.vision_service.preprocess_image(original_data)

        # Verify processed image is smaller
        processed_image = Image.open(io.BytesIO(processed_data))
        assert processed_image.size[0] <= 1024
        assert processed_image.size[1] <= 1024
        assert processed_image.format == "JPEG"

    def test_determine_confidence_level(self):
        """Test confidence level determination."""
        assert (
            self.vision_service._determine_confidence_level(0.9)
            == DiseaseConfidence.HIGH
        )
        assert (
            self.vision_service._determine_confidence_level(0.7)
            == DiseaseConfidence.MEDIUM
        )
        assert (
            self.vision_service._determine_confidence_level(0.5)
            == DiseaseConfidence.LOW
        )
        assert (
            self.vision_service._determine_confidence_level(0.3)
            == DiseaseConfidence.UNCERTAIN
        )

    def test_is_likely_crop_image(self):
        """Test crop image detection heuristic."""
        # Create a green image (likely crop)
        green_image = Image.new("RGB", (100, 100), "green")
        assert self.vision_service._is_likely_crop_image(green_image) is True

        # Create a blue image (unlikely crop)
        blue_image = Image.new("RGB", (100, 100), "blue")
        # This might still return True due to the simple heuristic, but test the method works
        result = self.vision_service._is_likely_crop_image(blue_image)
        assert isinstance(result, bool)

    def test_encode_image_base64(self):
        """Test base64 encoding."""
        image_data = self.create_test_image()

        encoded = self.vision_service._encode_image_base64(image_data)

        assert isinstance(encoded, str)
        # Verify it can be decoded back
        decoded = base64.b64decode(encoded)
        assert decoded == image_data

    def test_build_disease_analysis_prompt(self):
        """Test prompt building for disease analysis."""
        prompt = self.vision_service._build_disease_analysis_prompt(
            crop_hint="wheat",
            location_context="Punjab, India",
            additional_context="Leaves turning yellow",
        )

        assert "wheat" in prompt
        assert "Punjab, India" in prompt
        assert "Leaves turning yellow" in prompt
        assert "JSON format" in prompt

    @pytest.mark.asyncio
    async def test_identify_disease_invalid_image(self):
        """Test disease identification with invalid image."""
        invalid_data = b"not an image"

        with pytest.raises(ImageValidationError):
            await self.vision_service.identify_disease(invalid_data)

    @pytest.mark.asyncio
    @patch("app.services.vision_service.AsyncOpenAI")
    async def test_identify_disease_success(self, mock_openai):
        """Test successful disease identification."""
        # Mock OpenAI response
        mock_response = Mock()
        mock_response.choices = [Mock()]
        mock_response.choices[
            0
        ].message.content = """
        {
            "crop_type": "wheat",
            "diseases": [
                {
                    "disease_name": "wheat rust",
                    "confidence_score": 0.85,
                    "affected_parts": ["leaves"],
                    "severity": "moderate",
                    "description": "Orange pustules on leaf surface"
                }
            ],
            "treatment_recommendations": [
                {
                    "treatment_type": "chemical",
                    "active_ingredients": ["propiconazole"],
                    "dosage": "250ml per acre",
                    "application_method": "foliar spray",
                    "timing": "early morning",
                    "precautions": ["wear protective gear"],
                    "cost_estimate": "500-800 INR per acre"
                }
            ],
            "prevention_strategies": [
                "Use resistant varieties",
                "Proper crop rotation"
            ],
            "confidence_summary": "High confidence identification based on clear symptoms"
        }
        """

        mock_client = Mock()
        mock_client.chat.completions.create = AsyncMock(return_value=mock_response)
        mock_openai.return_value = mock_client

        # Create test image
        image_data = self.create_test_image()

        # Test disease identification
        analysis = await self.vision_service.identify_disease(image_data)

        assert analysis.crop_type == "wheat"
        assert len(analysis.diseases) == 1
        assert analysis.diseases[0].disease_name == "wheat rust"
        assert analysis.diseases[0].confidence_score == 0.85
        assert analysis.primary_disease is not None
        assert len(analysis.treatment_recommendations) == 1
        assert len(analysis.prevention_strategies) == 2

    @pytest.mark.asyncio
    @patch("app.services.vision_service.AsyncOpenAI")
    async def test_call_vision_model_error(self, mock_openai):
        """Test vision model API error handling."""
        # Mock OpenAI to raise an exception
        mock_client = Mock()
        mock_client.chat.completions.create = AsyncMock(
            side_effect=Exception("API Error")
        )
        mock_openai.return_value = mock_client

        image_data = self.create_test_image()
        base64_image = base64.b64encode(image_data).decode("utf-8")

        with pytest.raises(Exception):
            await self.vision_service._call_vision_model(base64_image, "test prompt")

    @pytest.mark.asyncio
    async def test_get_detailed_treatment_info(self):
        """Test getting detailed treatment information."""
        # This test would require mocking the RAG engine
        # For now, test that it doesn't crash
        try:
            result = await self.vision_service.get_detailed_treatment_info(
                disease_name="wheat rust", crop_type="wheat"
            )
            assert isinstance(result, dict)
            assert "detailed_treatment" in result
        except Exception:
            # Expected if RAG engine is not available
            pass

    @pytest.mark.asyncio
    @patch("app.services.vision_service.AsyncOpenAI")
    async def test_health_check(self, mock_openai):
        """Test vision service health check."""
        # Mock successful health check
        mock_response = Mock()
        mock_response.choices = [Mock()]
        mock_response.choices[0].message.content = "OK"

        mock_client = Mock()
        mock_client.chat.completions.create = AsyncMock(return_value=mock_response)
        mock_openai.return_value = mock_client

        health_status = await self.vision_service.health_check()

        assert health_status["status"] == "healthy"
        assert health_status["model"] == "gpt-4o"
        assert health_status["test_response"] == "OK"


class TestImageValidation:
    """Test cases for image validation functionality."""

    def test_supported_formats(self):
        """Test that all supported formats are recognized."""
        vision_svc = VisionService()

        # Test JPEG
        jpeg_data = self.create_test_image_data("JPEG")
        result = vision_svc.validate_image(jpeg_data)
        assert result.is_valid is True
        assert result.format == "jpeg"

        # Test PNG
        png_data = self.create_test_image_data("PNG")
        result = vision_svc.validate_image(png_data)
        assert result.is_valid is True
        assert result.format == "png"

    def create_test_image_data(self, format):
        """Helper to create test image data."""
        image = Image.new("RGB", (100, 100), "green")
        buffer = io.BytesIO()
        image.save(buffer, format=format)
        return buffer.getvalue()


@pytest.mark.integration
class TestVisionIntegration:
    """Integration tests for vision service (requires API keys)."""

    @pytest.mark.asyncio
    async def test_real_disease_identification(self):
        """Test with real API (requires OpenAI API key)."""
        # Skip if no API key
        from app.config import get_settings

        settings = get_settings()

        if not settings.openai_api_key:
            pytest.skip("OpenAI API key not configured")

        # Create a simple test image
        image = Image.new("RGB", (200, 200), "green")
        # Add some "disease-like" spots
        for i in range(10):
            for j in range(10):
                if (i + j) % 3 == 0:
                    image.putpixel((50 + i, 50 + j), (139, 69, 19))  # Brown spots

        buffer = io.BytesIO()
        image.save(buffer, format="JPEG")
        image_data = buffer.getvalue()

        try:
            analysis = await vision_service.identify_disease(
                image_data=image_data,
                crop_hint="test crop",
                additional_context="This is a test image with artificial spots",
            )

            assert analysis is not None
            assert analysis.image_id is not None
            assert analysis.processing_time > 0
            # Note: Results may vary as this is a synthetic test image

        except Exception as e:
            # Log the error but don't fail the test for API issues
            print(f"Integration test failed (expected with test image): {e}")


if __name__ == "__main__":
    pytest.main([__file__, "-v"])
