backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0

# Cache and session
redis==5.0.1

# File handling
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0

# HTTP client
httpx==0.25.2

# AI/ML APIs
openai==1.3.7
anthropic==0.7.8
chromadb==0.4.18
google-cloud-translate==3.12.1
# Geoapify API (Maps service) - uses standard HTTP client (aiohttp/httpx)

# Image processing
Pillow==10.1.0

# Communication
twilio==8.10.3
python-socketio==5.10.0
aiohttp==3.9.1

# Background tasks
celery==5.3.4

# Monitoring
prometheus-client==0.19.0

# Logging
python-json-logger==2.0.7

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
hypothesis==6.92.1